move user report processing logic form utils into the processor
